% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonctions_ocde.R
\name{filtrer_modalite}
\alias{filtrer_modalite}
\title{Filtrer un jeu de donnees selon une modalite}
\usage{
filtrer_modalite(data, variable_quali, modalite)
}
\arguments{
\item{data}{Un data frame.}

\item{variable_quali}{Le nom de la variable qualitative.}

\item{modalite}{La modalite a conserver.}
}
\value{
Un data frame filtre.
}
\description{
Cette fonction filtre un jeu de donnees a partir d'une variable qualitative
et d'une modalite choisie.
}
\examples{
filtrer_modalite(ocde, "continent", "Europe")
}
