% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonctions_ocde.R
\name{graphique_stat_par_groupe}
\alias{graphique_stat_par_groupe}
\title{Faire un graphique d'une statistique par groupe}
\usage{
graphique_stat_par_groupe(data, variable_quanti, variable_quali, stat = "mean")
}
\arguments{
\item{data}{Un data frame.}

\item{variable_quanti}{Le nom de la variable quantitative.}

\item{variable_quali}{Le nom de la variable qualitative de regroupement.}

\item{stat}{La statistique a calculer : \code{"mean"}, \code{"median"}, \code{"min"} ou \code{"max"}.}
}
\value{
Un objet ggplot.
}
\description{
Cette fonction produit un graphique en barres d'une statistique calculee
sur une variable quantitative selon les modalites d'une variable qualitative.
}
\examples{
graphique_stat_par_groupe(ocde, "esperance_vie_ans", "continent", "mean")
}
